requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pymongo
nltk
pandas
python-multipartpydantic-settings
pymysql
